Separated the student.service
