+++ frontend/client/src/services/student/student.service.ts
@@ -1,22 +1,14 @@
 /* Types */
 import { AuthToken } from "../../types/Auth";
 import { ColumnType } from "../../types/Columns";
-import { Label, LabelContent, LabelType } from "../../types/Label";
 import { Project } from "../../types/Project";
-import { Student, StudentStorageItem } from "../../types/Student";
+import { Student } from "../../types/Student";
 
 /* Components, services & etc. */
 import { callAPI, USE_SERVER } from "../api/api.service";
 import { defaultStudents } from "./default-students";
-import storage from "../storage/storage.service";
 
-// TODO: Get stuff from local storage
-export const checkStudentStatus = (student: Student): ColumnType => {
-    student;
-    return ColumnType.Applied;
-}
-
-// TODO: Request student statuses and add them to localstorage for caching
+// TODO: Request student status (if in team or not etc.) from ML
 export const getStudentStatus = (student: Student): ColumnType => {
     student;
 
@@ -36,38 +28,4 @@ export const getStudents = (projectID: Project["id"], token: AuthToken): Promise
     }
 
     return USE_SERVER ? callAPI<Student[]>(`/projects/${projectID}/students`, token).catch(errorHandler) : Promise.resolve(defaultStudents);
-}
-
-export const getStudentLabels = (studentId: Student["id"], labelType: LabelType): LabelContent[] => {
-    const saved = storage.get<StudentStorageItem>(`s${studentId}`);
-    return saved?.labels?.filter(label => label.isType === labelType).map(label => label.contains) ?? [];
-}
-
-const labelsAreSame = (a: Label, b: Label): boolean => a.isType === b.isType && a.contains.content === b.contains.content
-
-export const addLabelIfMissing = (studentId: Student["id"], label: Label): void => {
-    const itemName = `s${studentId}`;
-    const mapper = (old: StudentStorageItem) => {
-        const oldLabels = old.labels?.filter(lbl => !labelsAreSame(lbl, label)) ?? [];
-        return { ...old, labels: [ ...oldLabels, label] }
-    }
-
-    if (!storage.update<StudentStorageItem>(itemName, mapper)) {
-        storage.save<StudentStorageItem>({ labels: [label]}, itemName);
-    }
-}
-
-export const removeAllLabelsByType = (studentId: Student["id"], labelType: LabelType): void => {
-    const mapper = (old: StudentStorageItem) => {
-        return { ...old, labels: old.labels?.filter(lbl => lbl.isType !== labelType) ?? [] }
-    }
-    storage.update<StudentStorageItem>(`s${studentId}`, mapper);
-}
-
-export const removeLabelFromStudent = (studentId: Student["id"], label: Label): void => {
-    const mapper = (old: StudentStorageItem) => {
-        const oldLabels = old.labels?.filter(lbl => !labelsAreSame(lbl, label)) ?? [];
-        return { ...old, labels: oldLabels }
-    }
-    storage.update<StudentStorageItem>(`s${studentId}`, mapper);
 }
